feat: Base62UrlGenerator (#13)

+++ src/main/java/dev/saintho/mytly/generator/source/SourceGenerator.java
@@ -3,5 +3,5 @@ package dev.saintho.mytly.generator.source;
 import java.math.BigInteger;
 
 public interface SourceGenerator {
-	BigInteger generate(int base);
+	long generate(int base);
 }
